Use `react-height` for calculating height
The previous solution didn't take into account margins of children
of children (of children of children ...), and was therefore not
accurate with more complicated "expanded" panels.

Instead of trying to reinvent the wheel here, I suggest that we
start using `react-height` - which seemed to do the job just as
well. It also recomputes height whenever the children updates,
so a lot of what PR #6 suggests will be gained automatically.