Add RadioSwitch component
This commit adds a RadioSwitch component to the package, which
lets consumers create the FFE-specified binary choice component
"Radio switch".

In order to re-use a lot of the existing code, some internal
refactoring had to be done. Most of the structure and logic in
the RadioButton component is now found in the generic (and
"hidden" / unexposed) component RadioBase.

There are no breaking changes, but the RadioSwitch component is
added to the API, making this a minor version bump.