Add span element to space out .ffe-radio-switch
This commit adds some space between the radio switches. You can't
add it to the `.ffe-radio-switch` class without visually breaking
designs, and to be honest this is a React specific issue.

Since each radio button is wrapped in a div it typically isn't, it
doesn't get the spacing it should have gotten from being inline
elements. This `<span />` gives them that in the React world.