Allow setting the button type property
This commit refactors and changes the code to accept the default
`type` property. This leads to quite a few changes:

- Previous `type` property (used to specify button theme) is
renamed `buttonType`. This is a breaking change for consumers of the
`<Button />` component.
- `buttonType` property now verifies its value through `propTypes`.
- `type` property is passed as is to the `<button />` tag, and
verified to be either "submit", "button" or "reset". Defaults to
"submit".
- `onClick` is no longer a required prop, since type="submit"
triggers a `submit` event if put inside a form.

This commit also bumps the version to 3.0.0 because of its
breaking API changes.