Stop rendering children inside paragraph
This commit provides a solution for consumers who wants to send in
custom HTML to a message box.

Previously, children was put inside a `<p />` tag, which would
produce a `validateDOMNesting` warning if you sent in any block-
level elements (like other `<p />` tags, if you wanted two
paragraphs of texts instead of one) as children.

This commit renders `children` without this `<p />` tag, so that
the consumer can send in whatever it pleases.

If you want the current (old) behavior, it's still available by
sending your content in through the `content` property instead.

This is a potentially breaking change for consumers that are
expecting children to be wrapped in a `<p />` tag. Those consumers
can migrate to the new version either by adding the tag directly
or using the `content` property.