PPM-805. Add "p-pose-icon"
This icon was originally kept in the PPM project as it didn't have
a usecase outside of the private pension project.

With the introduction of AEM/Camunda to the project however, AEM
also needs access to this icon so they can use it in their task
reminder list which means we'd like to include it in FFE now.