FFE-121. Remove width styles from ffe-form
As discussed in FFE meeting we should avoid including layout
rules in our components and therefore having modifiers for
width embedded in a component is not optimal. What goes as
a "small" input in one application might not be "small" in
another.

As a result of this, all input controls now default to using
100% of their available width and it's up to the application
to decide what is available (or implement their own modifiers).

The defaults for display block/inline-block was kept because
they make sense and changing all of this twice in a row would
cause a lot of disruption.