Remove Gulp dependency, use svg-sprite directly
The Gulp 3.x branch had some pretty outdated dependencies
down the tree, which could cause trouble on Node 7.

Gulp was always a bit of an overkill for this package,
and it is about time we moved away from it.

This is a breaking change. The migration steps are described
in the changelog.