FFE-157 Bugfix and new feature
Fixes a bug where the list in icons.json would be ignored and all
icons included. Introduced in 7.0.0.

Adds a new option `projectIcons` to include icons external to
ffe-icons in the ffe-icons bundle. You could hack your way to this
functionality in versions before 7.0.0, but now it's an official
feature.