Formatting and sorting of things in alphabetical order
I know there isn't a common style guide for FFE, but I feel
most of the changes in this commit should be uncontroversial.

* Alphabetical order of props
* Formating of blocks in render
* Blank line at the end of files
* Some bling in the changelog

I also updated the example in the README to show the minimal usage
of the component rather than all props in the example since the
available props are listed in the next section anyway.