Improve InputGroup component
This commit adds support for the InputGroup component to receive
a function as a child instead of a component. This lets the
consumer handle the rendering completely, which in turn lets the
consumer send in several children to the InputGroup.

As a consequence of this, it's no longer possible to specify your
own id-prop. Although this is breaking, it was an anti-pattern to
begin with, and should not be heavily adopted.