Improve formatNumber
This commit adds the option to specify thousand separator and decimalmark to be used when formatting with formatNumber.

Default values is kept as it was, making it a non-breaking change. The implementation of `opts` argument is changed sligthly to allow for cases where only parts of the available options is specified, and still get the default values for the unspecified options.