Ensure id is a valid ID
This commit prefixes the id generated by the `InputGroup` component
with `input-` - which ensures it always is valid.

This change comes as a reaction to a transitive bug where some IDs
created by the UUID script started with a number. According to the
HTML spec, any ID has to start with a letter (or some special
characters - numbers not included).