Fix issue with wrapped reversed columns
This commit fixes a perceived bug in our API - when we add the
modifier `ffe-grid__row--reverse`, it only reverses the items per
row. The expected behavior would be to reverse the rows as well,
which this commit introduces.

Note that even though this is breaking some behavior, it's behavior
based on a bug - so if you're using row--reverse a lot in your code,
please test your code before upgrading.