fix(ffe-checkbox-react): Replace sdbm with uuid
sdbm uses a feature that is not supported by the version of UglifyJS
used by webpack 3.x (the unsigned right shift operator, >>>). Uglify
fails with an error message "Unexpected token: operator (>)" when trying
to uglify the function imported as sdbm. Replacing that dependency with
uuid fixes the issue, however the bug really lies with Uglify. It may
have been fixed in a later version - the one used by webpack is a bit
old. For now this will have to do.