chore(ffe-dropdown-react): Import ffe-dropdown-react
BREAKING CHANGE: Simplify the ffe-dropdown-react implementation.

This commit simplifies the component, and leaves more of the
compositional power to the consumer.

In order to migrate to this new major, the consumer should wrap
their existing component in `ffe-form-react`'s `InputGroup`
component. This will provide most of the functionality that this
major removes from `ffe-dropdown-react`. Please refer to the
documentation for new usage information.