fix(ffe-accordion-react): Remove "expandableContent" prop
BREAKING CHANGE: Replace "expandedContent" with children- & title- prop

This commit changes the structure of ffe-accordion-react by making
the expandable content of the accordion item the children, while the heading of the item
is now a new "title"-prop. This previous soulution was somewhat backwards,
while this change will be more the React way of doing things.