chore(ffe-spinner-react): Reimplement ffe-spinner-react
BREAKING CHANGE: This commit reimplements and simplifies the
ffe-spinner-react implementation. It no longer includes layout, which
requires the consumer to place it correctly.

The `<InlineSpinner />` and `<LargeSpinner />` are now available as
`<Spinner large={true|false} />`.

The previous `<Spinner />` implementation, which wrapped a large
spinner in a `ffe-content-container` class is now removed. To keep this
behavior, the consumer can create its own implementation of a
`<SpinnerWithLayout />` or something similar.