fix(ffe-account-selector-react): Explicit specify padding on "reset" button in account selector
This commit explicitly sets the horizontal padding on the reset button. Each of Google Chrome, Android WebView and iOS WebView have different values for padding on this button in their User Agent Stylesheet. On iOS this breaks the intended design.

The value used in this commit is the same value that is used on both Android and Desktop Chrome.