fix(ffe-tables): Fix overlapping media queries
The table media queries overlapped by one pixel, so at just the right
(or wrong) screen width styles for `md` and `lg` would both be applied.
The same with `lg` and `xlg`.

This version fixes the off-by-one in the media queries so only one is
active at the same time.