fix(ffe-buttons-react): Remove type propType check
This version removes the propType validation of the `type` property. It
makes sense if the element type is `button`, but now that consumers can
override the element type we run the risk of throwing propType errors
for valid values of `type` for element types we don't know about (custom
components for instance).