feat(ffe-buttons): Add inline-expand-button
This version introduces a new button meant to be used in-line at the end
of paragraphs. Clicking the button should expand more text below the
paragraph using an exapand animation (such as `react-collapse` for
instance. The button should always show a text that reflexts the
expanded state as well as a chevron icon with a toggle animation like
the accordion component.