feat(ffe-buttons-react): Inline expand button
This version introduces a React implementation of the inline expand
button. The component is meant to be used at the end of a paragraph
where other related paragraphs start off hidden. Clicking the button
should expand the hidden paragraphs. Composing this functionality is
left to the consumer. See InlineExpandButton.md for an example
implementation.