fix(ffe-details-list-react): Visual adjustments
After consulting with chief of design it was decided to use the
same stacking behaviour on md screens as we had on sm screens.
There will also be a discussion about the concept of having
6 columns on a medium screen since it forces us to either have
a lot of whitespace around a centered design or just the entire
horizontal space on what can still be a fairly large screen.
This stacking change is the quick fix while we wait for that
discussion to play out.

At the same time, the size of the labels was brought down from
`ffe-h5` to `ffe-h6` to better balance out the difference
between labels and the headings often used for sub-sections but
also to be more in line with the th-style of `ffe-tables`.

Two internal propTypes have had `.isRequired` removed from them
since the consumer is forced to create them without those props
and the library then later adds them. This should remove console
noise that the consumer would otherwise be unable to get rid of.