docs: Add line about messages -> changelog, example
Added an example of a great commit for breaking changes that includes a
diff and instructions for migrating in multiple scenarios.

I also change the link to some good practices for git commit messages to
a talk I really liked. My excuse is that the link to Medium failed to
render on Bitbucket and linked instead to
***REMOVED***