chore: Use suggested tool for running prettier
This commit replaces `precise-commits` with `pretty-quick`. Apart from
the obviously more clever name, `pretty-quick` has a few benefits to it:

- It's the suggested tool of Prettier, making it more heavily used and
  more maticiously maintained.
- It has really nicely formatted output
- It doesn't support patch-based running.

Now, this last part is the real reason why this change is warranted.
When you run `prettier` on a patch (and that patch alone), it doesn't
take into consideration the code around it. This led to mangled code
(that no longer even compiled), and post-commit unstaged changes. Nah.

Instead, `pretty-quick` formats any file that's been touched for a given
commit. Sure, it's going to format more code, leading to larger diffs,
but I'm confident the upsides outweight its downsides.

Note that this commit also removes prettier as a commit hook. This
pretty-quick tool needs to be tested manually before we can run it on
all commits.