infra: Add coverage reporting to CI build (#7)
This commit introduces a new test script for CI that includes coverage
reporting with Jest and Coveralls. It also uses Jests `--ci` option.
From the docs:

> When this option is provided, Jest will assume it is running in a CI
> environment. This changes the behavior when a new snapshot is
> encountered. Instead of the regular behavior of storing a new
> snapshot automatically, it will fail the test and require Jest to be
> run with `--updateSnapshot`.

The result from the coverage report is piped to a `coveralls` util
which takes care of submitting the coverage report to Coveralls and
further on to Github.

Travis will use this new CI script so we get Coveralls reporting on
pull requests.