fix: Improve experience when going to Komponenter
This commit introduces a new build system for the styleguide CSS. It
lets Webpack handle it by extracting the final CSS with
`ExtractTextWebpackPlugin` if the build is a production build. We use
`NODE_ENV` to determine if we're building PROD or not, and use
`cross-env` for compatability with other platforms.

Combined with `HtmlWebpackPlugin` which comes with React Styleguidist
this setup creates a separate CSS while which is linked in `head`. This
should fix the issue of the flash of unstyled content reported in #48.

The commit also brings changes to the Handlebars templates to improve
the loading experience while we're fetching the styleguide JavaScript
bundle, which can take a while.