docs: Reduce to one blank line between list items
The transformation from Markdown to HTML considers the items to be
separate lists if there is more than one blank line between the items.
The result would be fire ordered lists with one item, each being given
the number 1. by the browser.

This commit reduces the spacing between items to one blank line so they
all become part of the same ordered list, ranging from 1 to 5.