chore(ffe-icons-react): Upgrade to new ffe-icons
BREAKING CHANGE: This version migrates to the new `ffe-icons` which
replaced its method of generating the SVG sprite to no longer need
PhantomJS. This may or may not introduce visual regressions. From the
looks of it in the styleguide it should be a drop-in replacement,
however for safety this is marked as a breaking change.