fix(ffe-icons): Remove redundant XML tags and ns
When trying to make the SVG compatible with JSX it's helpful to have a
somewhat clean baseline. All the other icons are plain SVGs with only
the `xmlns` and maybe the `xmlns:svg` namespace definitions. These are
useful when using the SVGs standalone and can easliy be removed by the
`ffe-icons-react` build. Anything else should be removed.