fix(ffe-system-message): Apply its own layout
This commit adds the standard max width to this component's inner
container, which makes it layout agnostic. You no longer need to wrap
this component in any kind of grid or hawaii layout classes - it is
sized according to its container (with a correct maximum width).

This commit also adds some reset styles to the __close element class,
so that a more semantic `<button />` element can be used instead of
a `<div />`.