docs: Help developers use FFE for the first time
This commit fleshes out the documentation under Kom i gang for utviklere
by explaining a bit about the FFE packages and providing some links. For
convenience and flashiness the commit also introduces the Pakkeplukker
which is a widget to generate an `npm install` command for the most
popular packages with or without React.