fix: Visuell identitet a11y
There are a lot of id attributes in the SVGs that likely aren't used.
The SVGs themselves could also use a cleanup. But there are >100 IDs,
and the scope of this PR is already large enough.

This commit fixes some heading levels so they don't increase by more
than one.

It also makes the form examples more a11y compliant by adding labels and
groups.