fix(ffe-lists): Reset margin to allow using `dl`
For a11y and semantic reasons a `dd` and `dt` should only be direct
children of a `dl`. The multicolumn description list breaks this
guideline today, and without this margin reset the list becomes too
spacious when replacing the element wrapping the `dd+dt` pair with a
`dl`.

With this commit you can now do this:

```diff
-        <dl class="ffe-description-list-multicol">
+        <div class="ffe-description-list-multicol">
-            <div className="ffe-description-list-multicol__avoid-break">
+            <dl className="ffe-description-list-multicol__avoid-break">
                <dt>Navn</dt>
                <dd>Navn Navnesen</dd>
-            </div>
+            </dl>
-            <div className="ffe-description-list-multicol__avoid-break">
+            <dl className="ffe-description-list-multicol__avoid-break">
                <dt>Alder</dt>
                <dd>34</dd>
-            </div>
+            </dl>
             <!-- ect. -->
-        </dl>
+        </div>
```