fix(ffe-buttons): Properly align tertiary button anchors
We had a bug where the border and label of the tertiary button would be
rendered at the top and bottom of the button group if the button was an
anchor element (or !button, really). This commit adds the proper flex
attibute to make sure the content inside the container is vertically
centered. It also adjusts the CSS for the ::after element now that the
container takes care of the details.