chore: Add postcss for autoprefixing of webpack CSS
We had prefixing for the standalong ffe.css build, but not for
styleguidist itself. This commit adds postcss and tells it to add
prefixes so we get support for older IEs that some of our apps still
support.

We should probably move to use Webpack to also build ffe.css so we don't
have two separate pipelines for the two CSS files...