fix(ffe-buttons): fix ghost action transitions
This commit fixes the odd transitions of the ghost action button. As far
as I can tell, the only transition needed on the button label is the one
for transforms. Other transitions (background and color) is handled by
the actual button element.

I don't fully understand why this happened. It seems when you have an
element with `transition: all` and a child element with the same
transition styling, and you change color and background of the parent
element, the background of the parent will transition first, and then the
child element transition will run.
This is just my theory, so don't consider it the truth of css transitions.

Fixes #223