feat(ffe-form): Support hidden labels on checkbox
By adding the modifier `--hidden-label` the checkbox will be rendered
without the margins and paddings that are normally applied to textual
margins. This is usefull if you have a design that doesn't use visible
labels and you want to position the checkbox without having to work
around the margins and padding of a label that is not there.

When using this, it is important to remember to set the `aria-label`
attribute on the `<input>` element to facilitate for screenreaders.