feat(ffe-checkbox-react): Support hiddenLabel prop
Adds support for the prop `hiddenLabel` which can be used when
rendering a `Checkbox` without a visible label. Setting this prop
will remove all external padding and margin that would normally
surround the visible label and replace the `<label>` element
itself with a childless `<span>`.

When using this prop, remember to also use `aria-label` on the
`Checkbox` itself.