fix(ffe-form): fix placeholder color for searchable dropdown
This commit fixes the placeholder color the same way it's been fixed for
input fields. While searchable dropdown and account selector are
technically input fields, they're only styled with `.ffe-dropdown`. For
that reason we need to add this `::placeholder` element to this class
as well.

Note that this does not affect actual dropdowns, as they don't have a
placeholder pseudo element.

Fixes #240