fix(ffe-grid): Normalize padding and margin
When using the grid with certain elements (like `<ul>`) most
browsers will add a bit of margin and/or padding to the
element itself via user-agent styles. This fixes that by
explicitly setting those so that the `.ffe-grid` class behaves the
same way no matter what element it is put on.

Fixes #241