fix(ffe-core): Fix build issue
This commit fixes an issue where using the JS variables from ffe-core
broke certain builds (in my case, create-react-app). The reason for the
breakage was a trailing comma in a function argument list, which is an
ES6 feature.

This commit removes the trailing comma, and makes the code slightly
easier to read.