feat(ffe-core-react): Add `LinkText` component
Available as a named export from `ffe-core-react`, `LinkText` is a
wrapper around an `a` element with the proper classes. By default it
renders with an underline, but by setting the `underline` prop to
`false` the component will apply the `--no-underline` modifier.
Otherwise the component supports extending the `className` prop with
custom classes, and `...rest` props are spread directly on the anchor to
support things like `target` and `rel`.