docs: PoC of alternative information structure
We're fighting a bit with Styleguidist here... We want separate tabs
dedicated to usage, a summary of the key styles and states for the
components, and a tab for code examples. Right now most of the sections
are a mix that is difficult to navigate - if usage is documented at all.
This commit proposes a pattern using sections. One dedicated section for
usage and a summary of examples, and one section that lists the
individual components dedicated to documenting props and usage in code
like before.