docs: Add Bruk and Kode sections to :allthethings:
I chose to move many of the standalone packages that deal with form
inputs under the Skjema section. Unfortunately React Styleguidist
doesn't support sending in an array of globs, so we need to to that "by
hand" using glob to get a list of file names. Hence the somewhat ugly
components config for Skjema.

This commit is part one. Next comes moving existing docs we might have
to sensible locations.