feat(ffe-searchable-dropdown-react): Extended noMatch to allow for a set of fallback items.
This commit should be backwards compatible while allowing for a new set of items to be displayed for when the search results fail.

To use this feature, you need to replace the content of noMatch with an object that contains two fields `text` and `items`. `text` will be the text displayed at the top (same mechanism as the current functionality of noMatch) and `items` will be an array of items to display underneath.