feat(ffe-form): Restructure ffe-tooltip
BREAKING CHANGE: The base class `ffe-tooltip` includes
functionality for toggling the visibility of the tooltip.

The modified class `ffe-tooltip--open` controls the visibility of a
nested `ffe-tooltip__text` element. This modifier also deprecates the
modifier class `ffe-tooltip__icon--active`, which is no longer in use.

Important: The effect of the tooltip text `ffe-tooltip__text` appearing
and disappearing requires some scripting, because the underlying css
transition requires the element's `style.height` property set to an
absolute value.