Add section about peer-dependencies
I was thinking something like this can remedy the issue in #415. The only problem is... The depended URL does not show peerDepedencies and `ffe-form` is showing up because it has an optionalDependency so for this to work we'd need to use both.

Optional means that npm will try to install it but the process won't fail if it can't. I'm not sure what exactly happens if the project already depends on a different version of said package, that's maybe something we should test.

So in short, this can't be merged yet, we need to figure out a way to detect packages that will need updates. I just figured I'd get the ball rolling on this.