feat(ffe-cards-react): add new card components
This commit adds components for three new cards, as well as a card base
which can be used as a frame on which to build other cards. The commit
also adds some sub components. They are not supposed to be imported by
consumers, but instead are returned to a render function when such is
given to the cards as children. The cards that are added now will all
return all sub components, but we might want to restrict this for certain
cards at some point.