feat(ffe-cards): optional overflow ellipsis on title and left-align on text-card
This commit adds two modifiers to classes in ffe-cards:
1. ffe-text-card now has a modifier --left-align which will align content
to the left in the card. This will make the card work nicely together with
condensed icon cards.
2. ffe-card-component--title now requires the modifier --overflow-ellipsis
to limit the title to one line, and cut it off with ellipsis if it is too
long.
This commit also adds a few fixes:
* ffe-card-component--subtext now extends .ffe-small-text. The subtext was
too big, as it was always supposed to match the small text, and thus
extending it makes sense.
* ffe-icon-card__icon within a condensed icon card now has reduced
margins. The condensed version of icon card seemed to have too much air,
and this fix addresses that issue.
* ffe-image-card got its image max-height increased to 160px and its body
min-height reduced to 200px. Both adjustments are by 20px, so the overall
size of the card remains the same (unless your body exceeds 200px height).
* ffe-text-card got left and right padding increased from 20px to 30px.
This makes for a nicer look, especially when using the left-align modifier.
* ffe-text-card now has a min-height of 100px on screen sizes above
breakpoint medium. This is the same as the condensed icon card, and makes
sure they look uniform when used together.

BREAKING CHANGE: ffe-card-component--title now requires the modifier
--overflow-ellipsis to limit the title to one line, and cut it off with
ellipsis.