feat(ffe-cards-react): left align text card and toggle title wrapping
This commit adds two modifier properties to components in ffe-cards-react:
1. TextCard now has a bool property leftAlign which will align all text
content to the left in the card.
2. Render-prop component Title now has a bool prop overflowEllipsis which
will disable wrapping and hide overflow with ellipsis. This is true by
default, and thus this change is not breaking.
Other improvements to this package:
* TextCard, ImageCard and IconCard now has element in their prop types.
This is actually a property on CardBase, but we can help IDEs by informing
that these components accepts this property.

Please note this requires new major version 7.0.0 of ffe-cards.