feat(ffe-tables-react): Expandable table rows may be expanded as default
This version adds support for expanding a table row by default. The cells prop in TableRowExpandable may have a value defaultExpanded, that row is then expanded as default and scrolled to on mount.

This functionality is needed eg. when the expanded content is containg a link to another page, and that page has a back button. We then want to keep the position in the table, and have the option still expanded, when pressing the back button.