feat(ffe-form): Bottom padding of input controls
BREAKING CHANGE:
All FFE input controls that are contained within a `.ffe-input-group` or a `.ffe-fieldset` now by
default has added margin to the bottom. The space this creates is meant to be used up by a validation
message of type `.ffe-field-error-message` which now has a negative top margin once inside any of the
above mentioned containers. This means that if a single-line error message is added underneath an
input control, content further down on the page will not "jump" downwards.

If you don't want this behaviour, use the `--no-extra-margin` modifier on your `.ffe-fieldset`s and
`.ffe-input-group`s.

Fixes #381