chore: bump React to 16.9 to allow for using hooks etc
BREAKING CHANGE: peerDep of React is raised to 16.9. While this in itself
isn't a breaking change, it allows for the use of hooks in later releases.
Such changes would just be a minor version, or even a patch, but without
this major version bump it would result in a runtime error for consumers.