refactor(ffe-account-selector-react): remove deprecated lifecycle
BREAKING CHANGE: This commit removes the deprecated lifecycle method
`componentWillReceiveProps` from the `InputField` component. The component
used this lifecycle to keep the `value` prop and the internal `value`
state in sync. However this is a bad pattern, as there is no reason for
this component to keep an internal state of the field value.
The component _should_ be used as a controlled component, where the
consumer is responsible for keeping and providing the value. If you do this
already, you'll be fine with. However, if you've simply been listening for
changes through `onChange`, and not keeping your own state, this change will
break your UI.