fix(ffe-context-message-react): Allow custom header tag
This fix adds a new prop that allow users to decide what HTML element
the title in context message should be. It'll default to a div if
nothing is set.  Previously it would generate a <header> element, but
this is not semantically correct, and since this component can be placed
anywhere on a page, it did not make sense to hardcode a heading level.