fix(ffe-webfonts): Makes SpareBank1-font family opt-in
Due to download size the font files of the new font family `SpareBank1`
is not included anymore by default, but has to be imported by the
consumer. This is either done by importing `sb1-fonts.css` (fonts
inlined as data-uri) or `sb1-fonts.less` (fonts linked as `url()`) from
this package.