feat(ffe-searchable-dropdown-react): Allow selecting item programmatically
BREAKING CHANGE: Replaced `initialValue` with `selectedItem`

`selectedItem` is an OPTIONAL prop which enables you to programmatically decide which item
is displayed in the input field. When the user selects a different item, you should
update the `selectedItem` value to reflect this newly chosen item. If not, your props
and the internal state of SearchableDropdown will become out of sync.