fix(ffe-datepicker-react): make the datepicker work in shadow dom
We should always close an opened datepicker, whenever a click happens
outside the datepicker. To enable this, the datepicker registers a
global click handler on window, that checks the origin of every click
event.

Previously, the global click handler used Element.contains() to check if
the event target was contained in the components DOM tree. This does not
work if the component is rendered in a shadow DOM, since events
originating from a shadow DOM are retargeted, and thus the event target
is not contained in the components DOM tree. The result is that the same
click event that opens the datepicker, also closes the datepicker when
the event bubbles to window.

We make the datepicker work in shadow DOM by setting a flag on all click
events originating from the datepicker, and change the logic in the
global click handler to check for this flag instead of using
Element.contains().