feat(ffe-searchable-dropdown-react): Add support for tens of thousands of options
Using react-virtualzed we only render the list options that are dislayed to the user.
This enables us to go from at most a few thousand options to virtually an unlimited amount.
This implementation has been tested with 300000 options on Chrome.

This aims to solve the same issue that the highCapacity prop of AccountSelector was
introduced to solve in 93f21898c85988153091b15af7704df1d51ef105