fix(ffe-searchable-dropdown-react): make it work in shadow dom
Clicks outside the component should close the list of suggestions.
This is currently enabled by registering a global event handler on
document, that listens for "mousedown" and "focusin" event, which
closes the list of suggestions if the event originated outside the
component.

Previously, the global event handler used Element.contains() to check if
the event target was contained in the components DOM tree. This does not
work if the component is rendered in a shadow DOM, since events
originating from a shadow DOM are retargeted, and thus the event target
is not contained in the components DOM tree. The result is that the same
mousedown event that should choose a suggestion, instead ends up closing
the list of suggestions.

We make the searchable dropdown work in shadow DOM by setting a flag on
all mousedown and focusin events originating from the component, and
changing the logic in the global event handler to check for this flag.

NB: The "event flag" check will only work in react v17, due to changes
in event handling in react v17. This in turns means that this fix will
only have an effect in react v17, but shadow DOM is broken with
earlier react versions than v16 anyway.