chore: release action checks repository owner before running
The github action for releasing will now check name and owner of the
repository and only run if it is the original repository on
github/SpareBank1/designsystem. This will prevent the release action to
try to publish our packages from forks of this repository, which can
cause unexpected errors for the forker.