feat(ffe-datepicker-react): be smarter about year input
DatePicker assumed a year written as two digits is in the 21th century,
e.g. 010199 becomes 1. jan 2099. This can be an unwanted behaviour when
asking for past dates like birth dates.

With this change DatePicker takes the value of `maxDate` into
consideration when guessing the century. Setting `maxDate` to today's
date will place a two digit year within the hundred years before today,
which is a plausible assumption when asking for birth dates.